+++ src/funding/common.jsx
@@ -25,7 +25,8 @@ export type LogoOptions = {|
     fundingEligibility : FundingEligibilityType,
     onClick? : (event : MouseEvent, ...args: $ReadOnlyArray<mixed>) => void,
     onKeyPress? : (event : KeyboardEvent, ...args: $ReadOnlyArray<mixed>) => void,
-    nonce? : ?string
+    nonce? : ?string,
+    experiment : Experiment
 |};
 
 export type LabelOptions = {|
@@ -41,7 +42,6 @@ export type LabelOptions = {|
     onClick : (event : Event, ...args: $ReadOnlyArray<mixed>) => void,
     onKeyPress? : (event : KeyboardEvent, ...args: $ReadOnlyArray<mixed>) => void,
     layout : $Values<typeof BUTTON_LAYOUT>,
-    clientAccessToken : ?string,
     personalization : ?Personalization,
     nonce : ?string,
     tagline : ?boolean,
@@ -51,12 +51,13 @@ export type LabelOptions = {|
 export type WalletLabelOptions = {|
     locale : LocaleType,
     logoColor : $Values<typeof LOGO_COLOR>,
-    instrument : WalletInstrument,
+    instrument : ?WalletInstrument,
     content : ?ContentType,
     commit : boolean,
     experiment : Experiment,
     vault : boolean,
-    nonce? : ?string
+    nonce? : ?string,
+    textColor : $Values<typeof TEXT_COLOR>
 |};
 
 export type TagOptions = {|
@@ -65,6 +66,8 @@ export type TagOptions = {|
 |};
 
 export type FundingSourceConfig = {|
+    enabled : boolean,
+    automatic : boolean,
     shippingChange? : boolean,
     platforms : $ReadOnlyArray<$Values<typeof PLATFORM>>,
     layouts : $ReadOnlyArray<$Values<typeof BUTTON_LAYOUT>>,
@@ -118,6 +121,8 @@ export function BasicLabel({ logo, label, period, locale: { lang } } : LabelOpti
 }
 
 export const DEFAULT_FUNDING_CONFIG : FundingSourceConfig = {
+    enabled:   true,
+    automatic: true,
 
     layouts: [
         BUTTON_LAYOUT.VERTICAL