+++ src/funding/funding.js
@@ -2,7 +2,7 @@
 
 import type { FundingEligibilityType } from '@paypal/sdk-client/src';
 import { PLATFORM, FUNDING, COMPONENTS } from '@paypal/sdk-constants/src';
-import { SUPPORTED_FUNDING_SOURCES } from '@paypal/funding-components/src';
+import { values } from 'belter/src';
 
 import type { Wallet } from '../types';
 import { BUTTON_LAYOUT, BUTTON_FLOW } from '../constants';
@@ -10,9 +10,19 @@ import type { OnShippingChange } from '../ui/buttons/props';
 
 import { getFundingConfig } from './config';
 
-export function isFundingEligible(source : $Values<typeof FUNDING>, { layout, platform, fundingSource, fundingEligibility, components, onShippingChange, flow, wallet } :
-    {| layout? : $Values<typeof BUTTON_LAYOUT>, platform : $Values<typeof PLATFORM>, fundingSource : ?$Values<typeof FUNDING>, flow : $Values<typeof BUTTON_FLOW>,
-    fundingEligibility : FundingEligibilityType, components : $ReadOnlyArray<$Values<typeof COMPONENTS>>, onShippingChange : ?Function, wallet? : ?Wallet |}) : boolean {
+type IsFundingEligibleOptions = {|
+    layout? : $Values<typeof BUTTON_LAYOUT>,
+    platform : $Values<typeof PLATFORM>,
+    fundingSource : ?$Values<typeof FUNDING>,
+    flow : $Values<typeof BUTTON_FLOW>,
+    fundingEligibility : FundingEligibilityType,
+    components : $ReadOnlyArray<$Values<typeof COMPONENTS>>,
+    onShippingChange : ?Function,
+    wallet? : ?Wallet
+|};
+
+export function isFundingEligible(source : $Values<typeof FUNDING>,
+    { layout, platform, fundingSource, fundingEligibility, components, onShippingChange, flow, wallet } : IsFundingEligibleOptions) : boolean {
 
     if (!fundingEligibility[source] || !fundingEligibility[source].eligible) {
         return false;
@@ -24,12 +34,24 @@ export function isFundingEligible(source : $Values<typeof FUNDING>, { layout, pl
         return false;
     }
 
+    if (!fundingConfig.enabled) {
+        return false;
+    }
+
+    if (!fundingConfig.automatic && source !== fundingSource) {
+        return false;
+    }
+
     if (fundingConfig.eligible && !fundingConfig.eligible({ components, fundingSource, fundingEligibility, layout, wallet })) {
         return false;
     }
 
     if (layout && fundingConfig.layouts && fundingConfig.layouts.indexOf(layout) === -1) {
-        return false;
+        if (fundingSource && layout === BUTTON_LAYOUT.HORIZONTAL) {
+            // continue
+        } else {
+            return false;
+        }
     }
 
     if (fundingConfig.platforms && fundingConfig.platforms.indexOf(platform) === -1) {
@@ -56,17 +78,19 @@ export function determineEligibleFunding({ fundingSource, layout, platform, fund
         return [ fundingSource ];
     }
 
-    let eligibleFunding = SUPPORTED_FUNDING_SOURCES.filter(source =>
+    let eligibleFunding = values(FUNDING).filter(source =>
         isFundingEligible(source, { layout, platform, fundingSource, fundingEligibility, components, onShippingChange, flow, wallet }));
 
     if (layout === BUTTON_LAYOUT.HORIZONTAL) {
         eligibleFunding = eligibleFunding.slice(0, 2);
+    } else if (layout === BUTTON_LAYOUT.VERTICAL) {
+        eligibleFunding = eligibleFunding.slice(0, 6);
     }
 
     return eligibleFunding;
 }
 
-export function isVaultedFundingEligible({ wallet, onShippingChange } : {| wallet : ?Wallet, onShippingChange : ?OnShippingChange |}) : boolean {
+export function isWalletFundingEligible({ wallet, onShippingChange } : {| wallet : ?Wallet, onShippingChange : ?OnShippingChange |}) : boolean {
     if (!wallet) {
         return false;
     }