+++ src/marks/component.jsx
@@ -16,13 +16,20 @@ import { MarksElement } from './template';
 
 const DEFAULT_HEIGHT = 20;
 
-type MarksType = {|
+type MarksInstance = {|
     isEligible : () => boolean,
     render : (string | HTMLElement) => ZalgoPromise<void>
 |};
 
-export const getMarksComponent = memoize(() => {
-    function Marks({ fundingSource, onShippingChange } : {| fundingSource? : ?$Values<typeof FUNDING>, onShippingChange? : OnShippingChange |} = {}) : MarksType {
+type MarksProps = {|
+    fundingSource? : ?$Values<typeof FUNDING>,
+    onShippingChange? : OnShippingChange
+|};
+
+export type MarksComponent = (MarksProps) => MarksInstance;
+
+export const getMarksComponent : () => MarksComponent = memoize(() => {
+    function Marks({ fundingSource, onShippingChange } : MarksProps = {}) : MarksInstance {
 
         const height = DEFAULT_HEIGHT;
         const fundingEligibility = getFundingEligibility();
@@ -32,6 +39,7 @@ export const getMarksComponent = memoize(() => {
         const components = getComponents();
         const flow = BUTTON_FLOW.PURCHASE;
         const fundingSources = determineEligibleFunding({ fundingSource, fundingEligibility, components, platform, remembered, layout, flow });
+        const experiment = {};
 
         const isEligible = () => {
             if (!fundingSource) {
@@ -54,6 +62,7 @@ export const getMarksComponent = memoize(() => {
                                 fundingEligibility={ fundingEligibility }
                                 fundingSources={ fundingSources }
                                 height={ height }
+                                experiment={ experiment }
                             />
                         </div>
                     ).render(dom({ doc: document }))